[chore] move attributes func into separate file
No functional change, just moving it into its own file to clean up the tracer config since the meter provider will also use attributes.
